pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Daniyal <daniyal.mambetov.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
djoser = "^2.3.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
redis = "^5.2.0"
django-redis = "^5.4.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
swagger-ui-bundle = "^1.1.0"
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
